Use IO exceptions fluent API in serial reader

+++ src/Engine.Core/FileSerialization/Common/Exceptions/IO/EndOfFileOperationException.cs
@@ -6,20 +6,20 @@ internal sealed class EndOfFileOperationException : OperationException
 {
     private new const string MessageTemplate = $"{OperationException.MessageTemplate}; end of file reached. Attempted to {{0}} {{2}} byte(s) beyond the file length.";
 
-    public EndOfFileOperationException(OperationType operation, int bytesCount, int trailingBytesCount)
+    public EndOfFileOperationException(OperationType operation, int bytesCount, long trailingBytesCount)
         : base(operation, bytesCount, ComposeMessage(operation, bytesCount, trailingBytesCount))
     {
         TrailingBytesCount = trailingBytesCount;
     }
 
-    public EndOfFileOperationException(OperationType operation, int bytesCount, int trailingBytesCount, Exception inner)
+    public EndOfFileOperationException(OperationType operation, int bytesCount, long trailingBytesCount, Exception inner)
         : base(operation, bytesCount, ComposeMessage(operation, bytesCount, trailingBytesCount), inner)
     {
         TrailingBytesCount = trailingBytesCount;
     }
 
-    public int TrailingBytesCount { get; }
+    public long TrailingBytesCount { get; }
 
-    private static string ComposeMessage(OperationType operation, int bytesCount, int trailingBytesCount) =>
+    private static string ComposeMessage(OperationType operation, int bytesCount, long trailingBytesCount) =>
         string.Format(MessageTemplate, GetMessageComponent(operation), bytesCount, trailingBytesCount);
 }
